feat: Split page.tsx inot components
